Tags which are assigned to a workItem may only be deleted using the force

+++ Assignment3.Entities/TagRepository.cs
@@ -4,10 +4,12 @@ namespace Assignment3.Entities;
 public class TagRepository : ITagRepository
 {
     private readonly KanbanContext context;
+    private readonly TaskRepository taskRepository;
 
     public TagRepository(KanbanContext _context)
     {
         context = _context;
+        taskRepository = new TaskRepository(_context);
     }
 
     public (Response Response, int TagId) Create(TagCreateDTO tag)
@@ -44,6 +46,11 @@ public class TagRepository : ITagRepository
         if (entity.tasks.Count > 0 && !force)
         {
             return Response.Conflict;
+        }
+
+        var tasksWithTag = taskRepository.ReadAllByTag(tagId.ToString());
+        if(tasksWithTag.Count() > 0 && !force) {
+            return Response.BadRequest;
         }
 
         context.Tags.Remove(entity);
